Create tinycolony
The Tiny Colony Base Game allows players to start building a colony without requiring any NFTs. Players are able to enjoy the Construction and Management Simulator elements of Tiny Colony by building a basic set of cells, and interacting with the base game characters in the "Training Grounds". Beyond that, a plethora of NFTs are available for players who want to maximize their experience and participate in community events in the Tiny Colony Metaverse.